feat: Finally done with stage 1
