+++ src/core/command/deposit.ts
@@ -4,7 +4,7 @@ import { MyContext } from '../helpers';
 export const handleDeposit = async (ctx: CommandContext<MyContext>): Promise<void> => {
   if (ctx.session.loggedIn) {
     await ctx.reply('Input amount to deposit in ₦ (Naira)');
-    ctx.session.state = 'depositRequestInProgress';
+    ctx.session.route = 'depositRequestInProgress';
   } else {
     await ctx.reply('**Login Required** 🔒\n\nUse /login to access this feature.');
   }
