--- src/core/command/index.ts
@@ -1,58 +1,0 @@
-import { Request, Response } from 'express';
-import { Users, IUser } from '../models/users';
-import bcrypt from 'bcrypt';
-import { handleError } from '../helpers';
-
-export const signup = async (req: Request, res: Response): Promise<void> => {
-  try {
-    const name: string = req.body.name;
-    const email: string = req.body.email;
-    const password: string = req.body.password;
-
-    if (!name || !email || !password) {
-      return handleError(res, 400, 'All Fields are Mandatory');
-    }
-
-    if (await Users.findOne({ where: { email } })) {
-      return handleError(res, 400, 'User Already Exists. Login Instead');
-    } else {
-      const hashedPassword = await bcrypt.hash(password, 10);
-      const result = { name, email, password: hashedPassword };
-      const user = await Users.create(result);
-
-      if (user) {
-        res.status(201).send();
-        return;
-      } else {
-        return handleError(res, 400, 'Invalid Data sent during signup process');
-      }
-    }
-  } catch (err) {
-    console.error(err);
-    return handleError(res, 500, `${err}`);
-  }
-};
-
-export const login = async (req: Request, res: Response): Promise<void> => {
-  const email: string = req.body.email;
-  const password: string = req.body.password;
-
-  if (!email || !password) {
-    return handleError(res, 400, 'All Fields are Mandatory');
-  }
-  const user: IUser | null = await Users.findOne({ email });
-
-  if (user) {
-  } else {
-    return handleError(res, 401, 'Email or Password is invalid');
-  }
-};
-
-// export const webhook = async (req: Request, res: Response): Promise<void> => {
-//   try {
-//     res.sendStatus(200);
-//   } catch (err) {
-//     console.error(err);
-//     return handleError(res, 500, `${err}`);
-//   }
-// };
