+++ src/core/interfaces/index.ts
