+++ src/core/models/transactions.ts
@@ -1,5 +1,5 @@
 import mongoose, { Schema, Document } from 'mongoose';
-import { TransactionType, TransactionStatus, FileType } from '../interfaces/models';
+import { TransactionType, TransactionStatus, FileType } from '../interfaces';
 
 export interface ITransactions extends Document {
   user_id: Schema.Types.ObjectId;
