--- src/core/routes/index.ts
@@ -1,9 +1,0 @@
-import express from 'express';
-// import { asyncHandler } from '../middleware/asyncHandler';
-// import { signup, login, webhook } from '../controllers';
-
-const router = express.Router();
-
-// router.post('/webhook', asyncHandler(webhook));
-
-export default router;
