--- src/scaling/types/CustomContext.ts
@@ -1,6 +1,0 @@
-import { SessionData } from './SessionData';
-import type { Context, SessionFlavor } from 'grammy';
-
-type CustomContext = Context & SessionFlavor<SessionData>;
-
-export type { CustomContext };
