--- src/scaling/types/SessionData.ts
@@ -1,8 +1,0 @@
-interface SessionData {
-  route: string;
-
-  leftOperand: number;
-  rightOperand: number;
-}
-
-export { SessionData };
